Refactor settings nav into component
