+++ frontend/src/routes/mod.rs
@@ -84,7 +84,7 @@ pub fn switch(route: AppRoute) -> Html {
 pub fn switch_settings(route: SettingsRoute) -> Html {
     match route {
         SettingsRoute::Profile => html! {<Settings />},
-        SettingsRoute::Account { user_id } => html! {<Settings />},
+        SettingsRoute::Account { user_id: _ } => html! {<Settings />},
         SettingsRoute::Tickets => html! {<Settings />},
         SettingsRoute::NotFound => html! {<Redirect<AppRoute> to={AppRoute::NotFound}/>}
     }
